backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.0
blinker==1.9.0
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3

# Dependencias para autenticación
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
PyJWT==2.10.1
flask-jwt-extended==4.6.0

# Dependencias para MongoDB
pymongo==4.10.1
python-dotenv==1.0.1

# Futuras dependencias para el chatbot
# openai>=1.0.0
# langchain>=0.1.0
# langchain-mongodb>=0.1.0 